created utils dir in users app and move UserQuotaManager in the dir in quota_manager.py file
